+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.5.2 - 2020-01-25
+
+### Changed
+
+-   **[LSK]** drop `betanet` support (03e9003e, @faustbrian)
+
+## 2.5.1 - 2020-01-25
+
+### Changed
+
+-   Use in-memory password for encryption if no password was provided (a0ab5fa7, @faustbrian)
+
+## 2.5.0 - 2020-01-25
+
+### Added
+
+-   Export plugin logo from metadata (ebe77d4d, @luciorubeens)
+
+### Changed
+
+-   **[BREAKING]** Require explicit profile encryption (0911acfd, @faustbrian)
+
 ## 2.4.5 - 2020-01-22
 
 ### Fixed